+++ view/view_show.php
@@ -24,11 +24,13 @@
         <div class="main">
             <br><br>
         <table id="message_list" class="message_list">
-            <tr>
-                <?php  echo $user->UserName; ?>
-                <a href="post/edit/<?php echo $posts->PostId; ?>"><img src="lib/parsedown-1.7.3/edit.png" width="30" height="20"  alt=""/></a>
-                <a href="post/delete_confirm/<?php echo $posts->PostId; ?>"><img src="lib/parsedown-1.7.3/delete.png" width="30" height="20"  alt=""/></a><br><br>
-            </tr>    
+            <?php if ($user): ?>
+                <tr>
+                    <?php  echo $user->UserName; ?>
+                    <a href="post/edit/<?php echo $posts->PostId; ?>"><img src="lib/parsedown-1.7.3/edit.png" width="30" height="20"  alt=""/></a>
+                    <a href="post/delete_confirm/<?php echo $posts->PostId; ?>"><img src="lib/parsedown-1.7.3/delete.png" width="30" height="20"  alt=""/></a><br><br>
+                </tr> 
+             <?php endif; ?>
             <tr>
                 <td>
                     <table> 
@@ -43,16 +45,20 @@
                         <?php foreach ($listanswer as $row): ?>
                             <table> 
                                 <tr>
+                                    <?php if ($user): ?>
                                     <td>
-                                        <a href="vote/addvote"><img src="lib/parsedown-1.7.3/vote1.png" width="30" height="20"  alt=""/></a><br>
-                                        <a href="vote/devote"><img src="lib/parsedown-1.7.3/vote2.png" width="30" height="20" alt=""/></a><br>
-                                         <a href="post/refuseanswer"><img src="lib/parsedown-1.7.3/refuser.png" width="30" height="20"  alt=""/></a>
+                                        <a href="vote/add_vote"><img src="lib/parsedown-1.7.3/vote1.png" width="30" height="20"  alt=""/></a><br>
+                                        <a href="vote/de_vote"><img src="lib/parsedown-1.7.3/vote2.png" width="30" height="20" alt=""/></a><br>
+                                         <a href="post/refuse_answer/<?php echo $row->PostId; ?>"><img src="lib/parsedown-1.7.3/refuser.png" width="30" height="20"  alt=""/></a>
                                          <a href="post/accept_answer/<?php echo $row->PostId; ?>"><img src="lib/parsedown-1.7.3/accepte.png" width="30" height="20" alt=""/></a>
                                      </td>
+                                      <?php endif; ?>
                                      <td>
                                         <li><?php echo $row->Body; ?></li><br>
+                                        <?php if ($user): ?>
                                          <a href="post/edit/<?php echo $posts->PostId; ?>"><img src="lib/parsedown-1.7.3/edit.png" width="30" height="20"  alt=""/></a>
                                          <a href="post/delete_confirm/<?php echo $row->PostId; ?>"><img src="lib/parsedown-1.7.3/delete.png" width="30" height="20"  alt=""/></a><br>
+                                         <?php endif; ?>
                                          <p>asked <span><?php echo $row->Timestamp; ?></span>
                                           &nbsp by <?php echo $row->name($row->AuthorId); ?> &nbsp by &nbsp</p>
                                      </td>
